[CHORE] Updated the work page
