+++ src/content/products.md
@@ -1,19 +1,23 @@
 ---
-title: 💪 Products
-heading: Products
+title: 💪 Work
+heading: Work
 utmParams: ['portfolio']
 products:
   - name: caligator
     link: 'https://www.producthunt.com/posts/caligator'
     thumbnail: caligator.png
+    description: 'An open-source calculator & convertor that thinks like you'
   - name: lazy-git
     link: 'https://github.com/SarathSantoshDamaraju/lazyGit'
     thumbnail: lazygit.png
+    description: 'Cut-short long git commands'
   - name: do-away
     link: 'https://www.producthunt.com/posts/doaway-email'
     thumbnail: do-away.png
+    description: 'Disposable emails simplified.'
   - name: revealurl 
     link: 'https://www.producthunt.com/posts/revealurl'
     thumbnail: revealurl.png
+    description: 'No more clicking unknown short-URLs'
 ---
