+++ src/styles/base.scss
@@ -10,10 +10,11 @@
 
 // Components
 @import './navbar.scss';
+@import './card.scss';
 
 // Pages
 @import './journals.scss';
-@import './products.scss';
+@import './work.scss';
 
 
 // 3rd party
