+++ src/styles/work.scss
@@ -1,0 +1,6 @@
+.products {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: left;
+}
